Fix: Use lexer_name="auto" in write_file_plugin
This commit fixes a bug in the `write_file_plugin` where the `Syntax` class was being used incorrectly. The `get_lexer_for_filename` method does not exist on the `Syntax` class. This has been corrected to use the `lexer_name="auto"` argument in the `Syntax` constructor, which allows `rich` to automatically detect the lexer from the filename.

This commit also includes a minor update to the `.Q/project.md` file.